+++ packages/database/src/index.ts
@@ -15,5 +15,7 @@ export { createInvitesRepository } from "./repositories/invites-repository";
 export { createProjectsRepository } from "./repositories/projects-repository";
 export { createEpicsRepository } from "./repositories/epics-repository";
 export { createGamesRepository } from "./repositories/games-repository";
+export { createBuildsRepository } from "./repositories/builds-repository";
+export type { BuildWithDetails } from "./repositories/builds-repository";
 export { createKnowledgeDocumentsRepository } from "./repositories/knowledge-documents-repository";
 export { createSubmissionsRepository } from "./repositories/submissions-repository";